finalization: clippy 0 предупреждений (CI-гейт честный), доки недели синхронизированы
34 варнинга: механика авто-фиксом (конверсии/касты/derive/doc-цитаты),
мёртвые дубликаты app::save_jpeg/rgb24_to_nv12 удалены (живой конвертер
в capture), linux-only диагностика гейted cfg_attr(not(linux), allow)
вместо ложного dead на Windows, unwrap-after-is_some -> if let с
filter(), алиас TickDiag для L3, осознанные allow(too_many_arguments)
на рендер/сквозных прогонах. Тесты 69/69, clippy 0.

Доки: README (69 тестов, MSP over USB VCP как основной транспорт),
ROADMAP (фаза E закрыта 09-09 soak-критериями, D — стендовая часть),
CHANGELOG — запись о линт-проходке. Terminal-тест: bash -> ssh ->
боевые каналы (ui_sim: 60 fps, e2e 9.1 мс, fc online, armed=false).

+++ crates/detector/src/lib.rs
@@ -461,8 +461,8 @@ pub fn tiles_2x2(w: u32, h: u32) -> Vec<Tile> {
     if w < 64 || h < 64 {
         return vec![Tile { x: 0, y: 0, w, h }];
     }
-    let tile_w = ((w as u32 * 11) / 20).max(32); // ~55 %
-    let tile_h = ((h as u32 * 11) / 20).max(32);
+    let tile_w = ((w * 11) / 20).max(32); // ~55 %
+    let tile_h = ((h * 11) / 20).max(32);
     let tile_w = tile_w.min(w);
     let tile_h = tile_h.min(h);
     let stride_x = w - tile_w;
